Updated TF version to required
I have updated the terraform version to be required.  I suppose if I wanted to be really slick I could try to look for the .terraform-version file, but that is really just a remnant of building us tfenv, which maybe we do not need anymore?  Maybe people would us it locally